requirements.txt
pandas
scikit-learn
fastapi
uvicorn
nltk
pyyaml
joblib
matplotlib
